Support running on simulators with ICS and newer
Simulators for ICS+ don't seem to have any network location
provider. This change enables us to run on such devices without
crashing.

We'll still get weather data in the simulator because we hard code the
location there anyway.

If this would happen on a real device, the user would get to see an
error message (but no temperature updates).